issue #79 chore : move package plan -> search
